refactor(backend)(game): extract & move update draft into separate turn draft service
