+++ apps/backend/src/test/kotlin/at/se2group/backend/game/service/GameServiceGetTest.kt
@@ -17,7 +17,6 @@ import org.mockito.junit.jupiter.MockitoExtension
 import java.time.Instant
 import java.util.Optional
 import org.junit.jupiter.api.BeforeEach
-import at.se2group.backend.persistence.TurnDraftRepository
 
 @ExtendWith(MockitoExtension::class)
 class GameServiceGetTest {
@@ -27,12 +26,9 @@ class GameServiceGetTest {
 
     lateinit var gameService: GameService
 
-    @Mock
-    lateinit var turnDraftRepository: TurnDraftRepository
-
     @BeforeEach
     fun setup() {
-        gameService = GameService(gameRepository, turnDraftRepository)
+        gameService = GameService(gameRepository)
     }
 
     @Test
